Restructure folders[Fixes bug]
