--- API/.coverage
@@ -1,2 +1,0 @@
-[report]
-omit =
