--- API/.gitignore
@@ -1,3 +1,0 @@
-venv/
-__pycache__/
-/__pycache__/
