--- API/api/responses.py
@@ -1,31 +1,0 @@
-"""Responses for all the reqests made in the app.py"""
-registration_successful = {
-        'success': True,
-        'message': "User Registered Succesfully.",
-        }
-login_successful = {
-        'success': True,
-        'message': "Logedin Succesfully.",
-        }
-login_fail = {
-        'success': False,
-        'message': "Login Failed."
-        }
-auth_fail = {
-        'success': False,
-        'message': "You are not authorised to access this page."
-        }
-request_fail = {
-        'success': False,
-        'message': "Not a valid Request ID."
-        }
-create_request_fail = {
-        'success': False,
-        'message': "All fields required."
-        }
-create_request_successful = {
-        'success': True,
-        'message': "Your request was submitted successfully.",
-        }
-requests = [{'orderId': '2004', 'foodorder': 'bacBHHHon',
-            u'description': u'fresh', 'quantity': '3'}]
