--- API/api/try.py
@@ -1,10 +1,0 @@
-
-4
-5
-6
-7
-# Dictionary of strings and int
-myList = ['f','a','s']
-#always removes last element of the list 
-myList.pop(2)
-print (myList)
