+++ API/methods.py
@@ -1,0 +1,5 @@
+from api.views import app
+
+"""Runs the app"""
+if __name__ == '__main__':
+    app.run(debug = True)
