--- API/requirements.txt
@@ -1,26 +1,0 @@
-astroid==1.6.4
-atomicwrites==1.1.5
-attrs==18.1.0
-click==6.7
-coverage==4.5.1
-Flask==1.0.2
-Flask-JWT==0.3.2
-Flask-Testing==0.7.1
-isort==4.3.4
-itsdangerous==0.24
-Jinja2==2.10
-lazy-object-proxy==1.3.1
-MarkupSafe==1.0
-mccabe==0.6.1
-more-itertools==4.2.0
-nose2==0.7.4
-passlib==1.7.1
-pluggy==0.6.0
-py==1.5.3
-PyJWT==1.4.2
-pylint==1.9.1
-pytest==3.6.0
-six==1.11.0
-Werkzeug==0.14.1
-wrapt==1.10.11
-gunicorn==19.9.0
